requirements.txt
pyyaml==6.0.1
kubernetes==29.0.0
envparse==0.2.0
